requirement.txt
alembic==1.6.5
appdirs==1.4.4
attrs==21.2.0
black==21.7b0
bravado==11.0.3
bravado-core==5.17.0
catboost==0.26
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
cliff==3.8.0
cmaes==0.8.2
cmd2==2.1.2
colorama==0.4.4
colorlog==5.0.1
cycler==0.10.0
flake8==3.9.2
future==0.18.2
gitdb==4.0.7
GitPython==3.1.18
graphviz==0.17
greenlet==1.1.0
idna==3.2
isort-black-compat==4.3.211
joblib==1.0.1
jsonpointer==2.1
jsonref==0.2
jsonschema==3.2.0
kiwisolver==1.3.1
lightgbm==3.2.1
Mako==1.1.4
MarkupSafe==2.0.1
matplotlib==3.4.2
mccabe==0.6.1
monotonic==1.6
msgpack==1.0.2
mypy-extensions==0.4.3
neptune-client==0.10.2
neptune-lightgbm==0.9.13
neptune-optuna==0.9.14
numpy==1.21.1
oauthlib==3.1.1
optuna==2.8.0
packaging==21.0
pandas==1.3.0
pathspec==0.9.0
pbr==5.6.0
Pillow==8.3.1
plotly==5.1.0
prettytable==2.1.0
psutil==5.8.0
pycodestyle==2.7.0
pydocstyle==5.1.1
pyflakes==2.3.1
PyJWT==2.1.0
pylama==7.7.1
pyparsing==2.4.7
pyperclip==1.8.2
pyrsistent==0.18.0
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2021.1
PyYAML==5.4.1
regex==2021.7.6
requests==2.26.0
requests-oauthlib==1.3.0
rfc3987==1.3.8
scikit-learn==0.24.2
scikit-plot==0.3.7
scipy==1.7.0
seaborn==0.11.1
simplejson==3.17.3
six==1.16.0
sklearn==0.0
smmap==4.0.0
snowballstemmer==2.0.0
SQLAlchemy==1.4.22
stevedore==3.3.0
strict-rfc3339==0.7
swagger-spec-validator==2.7.3
tenacity==8.0.1
threadpoolctl==2.2.0
tomli==1.0.4
tqdm==4.61.2
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
webcolors==1.11.1
websocket-client==1.1.0
